anti-alchemy/requirements.txt
Flask
gunicorn
psycopg2-binary